🎨 Move ParseResult type definitions in src/parser/internal/parse-result.ts file
